Extracting Solution Events
